requirements.txt
google-genai
python-dotenv
pyannote.audio
openai-whisper
pydub
torch
torchaudio
watchdog
tqdm